requirements.txt
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
secure-smtplib==0.1.1
Werkzeug==3.1.3
WTForms==3.2.1
openpyxl==3.1.5
google-auth-oauthlib==1.2.1
google-api-python-client==2.160.0
google-auth==2.38.0